+++ ecmwf_models/era5/reshuffle.py
@@ -31,17 +31,14 @@ import sys
 import argparse
 import numpy as np
 
-from pygeogrids import BasicGrid
-
 from repurpose.img2ts import Img2Ts
 from ecmwf_models.era5.interface import ERA5NcDs, ERA5GrbDs
-from ecmwf_models.utils import mkdate, parse_filetype
+from ecmwf_models.grid import ERA_RegularImgGrid, ERA5_RegularImgLandGrid
+from ecmwf_models.utils import mkdate, parse_filetype, parse_product, str2bool
 from datetime import time, datetime
 
-
-
-def reshuffle(input_root, outputpath, startdate, enddate, variables,
-              h_steps=[0,6,12,18], mask_seapoints=False, imgbuffer=200):
+def reshuffle(input_root, outputpath, startdate, enddate, variables, product=None,
+              h_steps=[0,6,12,18], land_points=False, imgbuffer=50):
     """
     Reshuffle method applied to ERA images for conversion into netcdf time
     series format.
@@ -58,13 +55,17 @@ def reshuffle(input_root, outputpath, startdate, enddate, variables,
         End date, from which images are read and time series are generated.
     variables: list or str
         Variables to read from the passed images and convert into time series format.
+    product : str, optional (default: None)
+        Either era5 or era5-land, if None is passed we guess the product from the
+        downloaded image files.
     h_steps : list, optional (default: [0, 6, 12, 18]
         Hours at which images are read for each day and used for reshuffling,
         therefore this defines the sub-daily temporal resolution of the time series that
         are generated.
-    mask_seapoints: bool, optional (default: False)
-        Mask points over sea, replace them with nan.
-    imgbuffer: int, optional (default: 50)
+    land_points: bool, optional (default: False)
+        Reshuffle only land points. Uses the ERA5 land mask to create a land grid.
+        The land grid is fixed to 0.25*0.25 or 0.1*0.1 deg for now.
+    imgbuffer: int, optional (default: 200)
         How many images to read at once before writing time series. This number
         affects how many images are stored in memory and should be chosen according
         to the available amount of memory and the size of a single image.
@@ -74,22 +75,44 @@ def reshuffle(input_root, outputpath, startdate, enddate, variables,
         h_steps = [0,6,12,18]
 
     filetype = parse_filetype(input_root)
+    product = parse_product(input_root) if not product else product
+
+
+    if land_points:
+        if product == 'era5':
+            grid = ERA5_RegularImgLandGrid(res_lat=0.25, res_lon=0.25)
+        elif product == 'era5-land':
+            grid = ERA5_RegularImgLandGrid(res_lat=0.1, res_lon=0.1)
+        else:
+            raise NotImplementedError(product, 'Land grid not implemented for product.')
+        subgrid = grid
+    else:
+        if product == 'era5':
+            grid = ERA_RegularImgGrid(res_lat=0.25, res_lon=0.25)
+        elif product == 'era5-land':
+            grid = ERA_RegularImgGrid(res_lat=0.1, res_lon=0.1)
+        else:
+            raise NotImplementedError(product, 'Grid not implemented for product.')
+        subgrid = None
 
     if filetype == 'grib':
+        if land_points:
+            raise NotImplementedError('Reshuffling land points only implemented for netcdf files')
+
         input_dataset = ERA5GrbDs(root_path=input_root, parameter=variables,
-                                  subgrid=None, array_1D=True, h_steps=h_steps,
-                                  mask_seapoints=mask_seapoints)
+                                  subgrid=subgrid, array_1D=True, h_steps=h_steps,
+                                  product=product, mask_seapoints=False)
     elif filetype == 'netcdf':
         input_dataset = ERA5NcDs(root_path=input_root, parameter=variables,
-                                 subgrid=None, array_1D=True, h_steps=h_steps,
-                                 mask_seapoints=mask_seapoints)
+                                 subgrid=subgrid, array_1D=True, h_steps=h_steps,
+                                 product=product, mask_seapoints=False)
     else:
         raise Exception('Unknown file format')
 
     if not os.path.exists(outputpath):
         os.makedirs(outputpath)
 
-    global_attr = {'product': 'ERA5 (from {})'.format(filetype)}
+    global_attr = {'product': '{} (from {})'.format(product.upper(), filetype)}
 
     # get time series attributes from first day of data.
     first_date_time = datetime.combine(startdate.date(), time(h_steps[0], 0))
@@ -98,7 +121,6 @@ def reshuffle(input_root, outputpath, startdate, enddate, variables,
     data = input_dataset.read(first_date_time)
     ts_attributes = data.metadata
 
-    grid = BasicGrid(data.lon, data.lat)
 
     reshuffler = Img2Ts(input_dataset=input_dataset, outputpath=outputpath,
                         startdate=startdate, enddate=enddate, input_grid=grid,
@@ -124,7 +146,7 @@ def parse_args(args):
     """
 
     parser = argparse.ArgumentParser(
-        description="Convert downloaded ERA5 image data into time series format.")
+        description="Convert downloaded ERA5/ERA5-Land image data into time series format.")
     parser.add_argument("dataset_root",
                         help='Root of local filesystem where the image data is stored.')
     parser.add_argument("timeseries_root",
@@ -138,10 +160,8 @@ def parse_args(args):
                         help=("Short name of variables as stored in the images, which are reshuffled. "
                               "See documentation on image download for resp. ERA products, "
                               "for more information on variable names of the product. "))
-    parser.add_argument("--mask_seapoints", type=bool, default=False,
-                        help=("Replace points over water with nan. This option needs the "
-                              "LandSeaMask (lsm) variable in the image data (will use mask from first available file). "
-                              "To use a dynamic LSM, reshuffle the LSM variable to time series."))
+    parser.add_argument("--land_points", type=str2bool, default='False',
+                        help=("Store only time series for points that are over land"))
     parser.add_argument("--h_steps", type=int, default=None, nargs='+',
                         help=("Time steps (full hours) of images that will be reshuffled (must be in the images). "
                               "By default 6H images (starting at 0:00 UTC) will be reshuffled."))
@@ -151,7 +171,7 @@ def parse_args(args):
                               "system and the size of a single image."))
     args = parser.parse_args(args)
 
-    print("Converting data from {} to {} into {}."
+    print("Converting ERA5/ERA5-Land data from {} to {} into {}."
           .format(args.start.isoformat(), args.end.isoformat(), args.timeseries_root))
 
     return args
@@ -165,7 +185,7 @@ def main(args):
               startdate=args.start,
               enddate=args.end,
               variables=args.variables,
-              mask_seapoints=args.mask_seapoints,
+              land_points=args.land_points,
               h_steps=args.h_steps,
               imgbuffer=args.imgbuffer)