+++ tests/test_grid.py
@@ -21,11 +21,12 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-'''
+"""
 Tests for grid generation
-'''
+"""
 
-from ecmwf_models.grid import ERA_RegularImgGrid, get_grid_resolution, ERA_IrregularImgGrid
+from ecmwf_models.grid import ERA_RegularImgGrid, get_grid_resolution, \
+    ERA_IrregularImgGrid, ERA5_RegularImgLandGrid
 import numpy as np
 
 def test_ERA_regular_grid():
@@ -34,7 +35,6 @@ def test_ERA_regular_grid():
     assert np.unique(reg_grid.activearrlon).size == 1200
     assert get_grid_resolution(reg_grid.activearrlat,
                                reg_grid.activearrlon) == (0.3, 0.3)
-
 
 def test_ERA_irregular_grid():
     # we test this with a regular grid, because it's easier
@@ -45,3 +45,18 @@ def test_ERA_irregular_grid():
     grid = ERA_IrregularImgGrid(lons, lats)
 
     assert grid == ERA_RegularImgGrid(1.,1.)
+
+def test_ERA5_landgrid_025():
+    grid = ERA5_RegularImgLandGrid(0.25, 0.25) # 0.25*0.25
+    assert grid.get_grid_points()[0].size == 244450
+    assert grid.find_nearest_gpi(16.375, 48.125)[0] == 240545
+    assert grid.gpi2cell(240545) == 1431
+
+
+def test_ERA5_landgrid_01():
+    grid = ERA5_RegularImgLandGrid(0.1, 0.1) # 0.1*0.1
+    assert grid.get_grid_points()[0].size == 1544191
+    assert grid.find_nearest_gpi(16.375, 48.125)[0] == 1508564
+    np.testing.assert_almost_equal(grid.gpi2lonlat(1508564)[0], 16.4)
+    np.testing.assert_almost_equal(grid.gpi2lonlat(1508564)[1], 48.1)
+    assert grid.gpi2cell(1508564) == 1431